history has a Invalid read of 16 when it uses gnl to push to double ptr

+++ srcs/main.c
@@ -28,8 +28,8 @@ static struct termios	ft_init_raw(void)
 		exit(1);
 	}
 	raw = orig_termios;
-	raw.c_lflag &= ~(size_t)(ICANON | ECHO | IEXTEN | ISIG);
-	raw.c_iflag &= ~(size_t)(IXON | BRKINT);
+	raw.c_lflag &= ~(ICANON | ECHO | IEXTEN | ISIG);
+	raw.c_iflag &= ~(IXON | BRKINT);
 	raw.c_cc[VMIN] = 1;
 	raw.c_cc[VTIME] = 0;
 	if (tcsetattr(STDIN_FILENO, TCSAFLUSH, &raw) == -1)
